Add socketAPI custom hook and use kline
