+++ src/Pages/Portfolio/Spline.js
@@ -1,4 +1,4 @@
-import React, { useEffect } from "react";
+import React from "react";
 import HighchartsReact from "highcharts-react-official";
 import Highcharts from "highcharts/highstock";
 import PropTypes from "prop-types";
@@ -7,7 +7,7 @@ import useCallBinanceAPI from "../../Hooks/useCallBinanceAPI";
 const Spline = (props) => {
   const { symbol } = props;
 
-  const [options, setOptions] = useCallBinanceAPI(
+  const [options] = useCallBinanceAPI(
     symbol,
     {
       plotOptions: {
@@ -45,31 +45,6 @@ const Spline = (props) => {
     },
     "spline"
   );
-
-  useEffect(() => {
-    const socket = new WebSocket(
-      `wss://stream.binance.com:9443/ws/${symbol}@kline_1m`
-    );
-    socket.onmessage = (event) => {
-      const newKLineData = [];
-      const data = JSON.parse(event.data);
-      newKLineData.push(data.k.t, (Number(data.k.h) + Number(data.k.l)) / 2);
-      setOptions((op) => {
-        const newOptions = { ...op };
-        newOptions.series[0].data = [
-          ...newOptions.series[0].data,
-          newKLineData,
-        ];
-        return newOptions;
-      });
-    };
-
-    return () => socket.close();
-  }, [setOptions, symbol]);
-
-  if (symbol === "") {
-    return null;
-  }
 
   return <HighchartsReact highcharts={Highcharts} options={options} />;
 };
